Implementação Parcial Tela Login Pt 3 (Logar/Deslogar)

+++ DTO/Conta.cs
@@ -12,8 +12,8 @@ namespace DTO
             public int IdCliente { get; set; }
             public string StatusConta { get; set; }
             public string TipoConta { get; set; }
-            public double Saldo { get; set; }
-            public double Limite { get; set; }
+            public decimal Saldo { get; set; }
+            public decimal Limite { get; set; }
             public DateTime DataAbertura { get; set; }
             public DateTime DataEncerramento { get; set; }           
             public string SenhaConta { get; set; }
